refactor: move validation funcs to own file
